refactor and add tests for van gogh case
