+++ package.json
@@ -10,12 +10,15 @@
     "node": "^20.6.0"
   },
   "devDependencies": {
+    "@jest/globals": "^29.6.4",
     "@tsconfig/node18": "^18.2.1",
+    "@types/jest": "^29.5.4",
     "@types/node": "^20.5.9",
     "husky": "^8.0.3",
     "jest": "^29.6.4",
     "lint-staged": "^14.0.1",
     "prettier": "^3.0.3",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   },
@@ -24,6 +27,7 @@
   },
   "scripts": {
     "start": "ts-node src/index.ts",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "test": "jest"
   }
 }
